Fix late binding
flake8-bugbear complains of "B023 Function definition does not bind loop
variable 'u'."

This is not really a bug, because the lambda only lasts as long as a
single iteration of the iterator. Fix the warning by passing the graph
as a parameter to the key lambda, which avoids binding the loop var.

Also refactor by moving _get_shortest_edges to own function. Too
hard to read otherwise.